feat: add test case for warehouse, mould and metal conversion in the workflow
