+++ .github/workflows/core-tests.yml
@@ -54,3 +54,4 @@ jobs:
           bench --site test_site run-tests --app jewellery_erpnext --doctype "Department IR"
           bench --site test_site run-tests --app jewellery_erpnext --doctype "Employee IR"
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_employee_ir_loss_baseline
+          bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_warehouse_stock_entry
