+++ jewellery_erpnext/jewellery_erpnext/tests/test_tree_casting.py
@@ -431,9 +431,9 @@ def _recv_eir(rows, typ="Receive", loss_rows=None):
 		type=typ,
 		manually_book_loss_details=[
 			SimpleNamespace(
-				variant_of="M", manufacturing_work_order=m, proportionally_loss=loss
+				variant_of="M", manufacturing_work_order=m, proportionally_loss=n
 			)
-			for m, loss in (loss_rows or [])
+			for m, n in (loss_rows or [])
 		],
 		employee_loss_details=[],
 		employee_ir_operations=[
@@ -886,11 +886,12 @@ class TestGetCastingGroupOperations(IntegrationTestCase):
 class _FakeSE:
 	"""Captures the Stock Entry the helper builds without touching the DB."""
 
-	def __init__(self):
+	def __init__(self, index=0):
 		self.items = []
 		self.name = None
 		self.flags = SimpleNamespace()
 		self.submitted = False
+		self._index = index
 
 	def append(self, _table, row):
 		child = SimpleNamespace(**row)
@@ -901,10 +902,16 @@ class _FakeSE:
 		return getattr(self, key, default)
 
 	def insert(self, *a, **k):
-		self.name = "SE-TREE-TEST-0001"
+		self.name = f"SE-TREE-TEST-{self._index:04d}"
 
 	def submit(self, *a, **k):
 		self.submitted = True
+
+
+class _RunResult(list):
+	"""The list of FakeSEs a helper created (in build order), plus ``.value`` = its return."""
+
+	value = None
 
 
 def _new_tree(employee_ir=None, material_details=None):
@@ -928,11 +935,22 @@ def _run(
 	msl_wh="EMP-MSL",
 	rm_wh="DEPT-RM",
 	scrap_wh="DEPT-SCRAP",
+	loss_item="GOLD-18KT-ML",
 ):
-	"""Call an op helper with persistence + warehouse resolution mocked. Returns the FakeSE."""
-	fake = _FakeSE()
+	"""Call an op helper with persistence + warehouse/loss-item resolution mocked.
+
+	Returns a _RunResult (list of every FakeSE created, in build order — receive posts the
+	received transfer first, then the loss Repack); ``.value`` holds the helper's return.
+	"""
+	fakes = _RunResult()
+
+	def _mint(*_a, **_k):
+		fake = _FakeSE(index=len(fakes))
+		fakes.append(fake)
+		return fake
+
 	with (
-		patch.object(tse.frappe, "new_doc", return_value=fake),
+		patch.object(tse.frappe, "new_doc", side_effect=_mint),
 		patch.object(tse.frappe, "has_permission", return_value=True),
 		patch.object(tse.frappe, "get_precision", return_value=3),
 		patch.object(tse, "_apply_fifo_batches_to_stock_entry"),
@@ -942,10 +960,10 @@ def _run(
 		patch.object(tse, "_resolve_msl_warehouse", return_value=msl_wh),
 		patch.object(tse, "_get_department_rm_warehouse", return_value=rm_wh),
 		patch.object(tse, "_resolve_scrap_warehouse", return_value=scrap_wh),
-		patch.object(tse, "_resolve_tree_loss_item", return_value="ML-LOSS-ITEM"),
+		patch.object(tse, "_resolve_tree_loss_item", return_value=loss_item),
 	):
-		fn(*args)
-	return fake
+		fakes.value = fn(*args)
+	return fakes
 
 
 # ---------------------------------------------------------------------------
@@ -954,7 +972,7 @@ def _run(
 class TestIssueMaterial(IntegrationTestCase):
 	def test_issue_builds_material_transfer_source_to_msl(self):
 		tree = _new_tree()
-		fake = _run(tse.issue_material, tree, "GOLD-18KT", 5.0)
+		fake = _run(tse.issue_material, tree, "GOLD-18KT", 5.0)[0]
 
 		# Plain Material Transfer (NOT WORK ORDER) => ledger-invisible, no reservation.
 		self.assertEqual(fake.stock_entry_type, "Material Transfer")
@@ -1014,33 +1032,46 @@ class TestReceiveMaterial(IntegrationTestCase):
 			]
 		)
 
-	def test_receive_builds_two_legs(self):
+	def test_receive_builds_transfer_and_loss_repack(self):
 		tree = self._issued_tree()
-		fake = _run(
+		fakes = _run(
 			tse.receive_material,
 			tree,
 			[{"item_code": "GOLD-18KT", "receive_qty": 6.0, "loss_qty": 1.0}],
 		)
-		# Both recv and loss SEs share the same _FakeSE stub (frappe.new_doc returns it
-		# twice). The stub accumulates all items: 1 from the Material Transfer recv SE
-		# + 2 from the Process Loss repack (consume + produce pair).
-		self.assertTrue(fake.submitted)
-		self.assertEqual(len(fake.items), 3)
-		received = fake.items[0]
-		# received leg: MSL -> Dept RM
+		# Two SEs: [0] received transfer, [1] loss Repack. Return value lists both names.
+		self.assertEqual(len(fakes), 2)
+		se_recv, se_loss = fakes
+		self.assertEqual(fakes.value, [se_recv.name, se_loss.name])
+
+		# Received leg: plain Material Transfer (standalone), MSL -> Dept RM, same metal item.
+		self.assertEqual(se_recv.stock_entry_type, "Material Transfer")
+		self.assertEqual(len(se_recv.items), 1)
+		received = se_recv.items[0]
+		self.assertEqual(received.item_code, "GOLD-18KT")
 		self.assertEqual(
 			(received.s_warehouse, received.t_warehouse), ("EMP-MSL", "DEPT-RM")
 		)
 		self.assertEqual(received.qty, 6.0)
-		# loss repack: consume metal at MSL (no target) + produce ML variant at Scrap (no source)
-		consume = fake.items[1]
-		produce = fake.items[2]
-		self.assertEqual(consume.s_warehouse, "EMP-MSL")
-		self.assertIsNone(consume.t_warehouse)
+
+		# Loss leg: Process Loss Repack — consume metal @ MSL, produce ML variant @ Scrap.
+		self.assertEqual(se_loss.stock_entry_type, "Process Loss")
+		self.assertEqual(se_loss.purpose, "Repack")
+		self.assertEqual(se_loss.auto_created, 1)
+		self.assertEqual(se_loss.custom_tree_number, tree.name)
+		self.assertEqual(len(se_loss.items), 2)
+		consume, produce = se_loss.items
+		self.assertEqual(consume.item_code, "GOLD-18KT")
+		self.assertEqual((consume.s_warehouse, consume.t_warehouse), ("EMP-MSL", None))
 		self.assertEqual(consume.qty, 1.0)
-		self.assertIsNone(produce.s_warehouse)
-		self.assertEqual(produce.t_warehouse, "DEPT-SCRAP")
+		# Produce row is the resolved ML loss variant, written off into Scrap.
+		self.assertEqual(produce.item_code, "GOLD-18KT-ML")
+		self.assertEqual(
+			(produce.s_warehouse, produce.t_warehouse), (None, "DEPT-SCRAP")
+		)
 		self.assertEqual(produce.qty, 1.0)
+		self.assertEqual(produce.is_finished_item, 1)
+		self.assertEqual(produce.set_basic_rate_manually, 1)
 
 	def test_receive_partial_then_full_status(self):
 		tree = self._issued_tree(10.0)
@@ -1066,21 +1097,22 @@ class TestReceiveMaterial(IntegrationTestCase):
 
 	def test_receive_only_loss_leg_when_no_receive(self):
 		tree = self._issued_tree(4.0)
-		fake = _run(
+		fakes = _run(
 			tse.receive_material,
 			tree,
 			[{"item_code": "GOLD-18KT", "receive_qty": 0, "loss_qty": 4.0}],
 		)
-		# Only the loss repack SE is posted (consume + produce = 2 items).
-		self.assertEqual(len(fake.items), 2)
-		consume = fake.items[0]
-		produce = fake.items[1]
-		self.assertEqual(consume.s_warehouse, "EMP-MSL")
-		self.assertIsNone(consume.t_warehouse)
-		self.assertEqual(consume.qty, 4.0)
-		self.assertIsNone(produce.s_warehouse)
-		self.assertEqual(produce.t_warehouse, "DEPT-SCRAP")
-		self.assertEqual(produce.qty, 4.0)
+		# Loss-only receive: a single Process Loss Repack SE (no transfer leg).
+		self.assertEqual(len(fakes), 1)
+		se_loss = fakes[0]
+		self.assertEqual(se_loss.stock_entry_type, "Process Loss")
+		self.assertEqual(se_loss.purpose, "Repack")
+		consume, produce = se_loss.items
+		self.assertEqual((consume.s_warehouse, consume.t_warehouse), ("EMP-MSL", None))
+		self.assertEqual(produce.item_code, "GOLD-18KT-ML")
+		self.assertEqual(
+			(produce.s_warehouse, produce.t_warehouse), (None, "DEPT-SCRAP")
+		)
 		self.assertEqual(tree.material_details[0].pending_qty, 0.0)
 		self.assertEqual(tree.status, "Received")
 
@@ -1116,59 +1148,115 @@ class TestReceiveMaterial(IntegrationTestCase):
 # Casting (employee_ir-seeded) tree buttons + resolvers
 # ---------------------------------------------------------------------------
 class TestCastingTreeButtons(IntegrationTestCase):
-	"""Casting trees now support the buttons: Issue posts a physical SE and owns issue_qty;
-	Receive is RECORD-ONLY (no SE — the Employee IR moves the physical metal) and auto-books
-	the remaining pending as dust, capped at the issued qty."""
+	"""Casting trees: Issue posts a `Material Transfer (MAIN SLIP)` SE and owns issue_qty; the
+	Employee IR Receive books the cast output, and the Receive button returns the post-cast
+	leftover (bounded by the pending cap, so it can never re-receive the EIR-booked qty)."""
 
-	def test_casting_issue_posts_se_and_increments_issue_qty(self):
+	def test_casting_issue_posts_main_slip_se(self):
 		tree = _new_tree(employee_ir="EIR-CASTING-0001")
-		fake = _run(tse.issue_material, tree, "GOLD-18KT", 5.0)
+		fake = _run(tse.issue_material, tree, "GOLD-18KT", 5.0)[0]
+		# Casting Issue is relabelled as MAIN SLIP (still ledger-invisible).
 		self.assertEqual(fake.stock_entry_type, "Material Transfer (MAIN SLIP)")
 		self.assertTrue(fake.submitted)
+		self.assertEqual(fake.items[0].s_warehouse, "SRC-MFG")
+		self.assertEqual(fake.items[0].t_warehouse, "EMP-MSL")
 		self.assertEqual(tree.material_details[0].issue_qty, 5.0)
 
-	def test_casting_receive_posts_se_for_casting_tree(self):
-		tree = _new_tree(
-			employee_ir="EIR-CASTING-0001",
-			material_details=[
-				{
-					"item_code": "GOLD-18KT",
-					"issue_qty": 10,
-					"receive_qty": 0,
-					"loss_qty": 0,
-					"pending_qty": 10,
-				}
-			],
-		)
-		fake = _run(
-			tse.receive_material, tree, [{"item_code": "GOLD-18KT", "receive_qty": 6.0}]
-		)
-		# Casting tree receive posts a real SE (Material Transfer MAIN SLIP).
-		self.assertTrue(fake.submitted)
-		self.assertEqual(len(fake.items), 1)
-		md = tree.material_details[0]
-		self.assertEqual(md.receive_qty, 6.0)
-		self.assertEqual(md.pending_qty, 4.0)
+	def test_standalone_issue_stays_plain_material_transfer(self):
+		tree = _new_tree()  # standalone (employee_ir empty)
+		fake = _run(tse.issue_material, tree, "GOLD-18KT", 5.0)[0]
+		self.assertEqual(fake.stock_entry_type, "Material Transfer")
 
-	def test_casting_receive_caps_at_issued(self):
-		tree = _new_tree(
+	def _casting_tree(self, issue=3.0, receive=2.0, loss=0.0):
+		# A casting tree whose Employee IR Receive has already booked `receive` (the cast output),
+		# leaving `issue - receive - loss` as the returnable leftover still sitting in MSL.
+		return _new_tree(
 			employee_ir="EIR-CASTING-0001",
 			material_details=[
 				{
 					"item_code": "GOLD-18KT",
-					"issue_qty": 10,
-					"receive_qty": 0,
-					"loss_qty": 0,
-					"pending_qty": 10,
+					"issue_qty": issue,
+					"receive_qty": receive,
+					"loss_qty": loss,
+					"pending_qty": issue - receive - loss,
 				}
 			],
 		)
+
+	def test_casting_receive_no_longer_throws(self):
+		# Casting trees return the post-cast leftover via the tree button as a MAIN SLIP transfer.
+		tree = self._casting_tree(issue=3.0, receive=2.0)  # EIR booked 2 -> pending 1
+		fake = _run(
+			tse.receive_material, tree, [{"item_code": "GOLD-18KT", "receive_qty": 1.0}]
+		)[0]
+		self.assertEqual(fake.stock_entry_type, "Material Transfer (MAIN SLIP)")
+		self.assertEqual(len(fake.items), 1)
+		self.assertEqual(
+			(fake.items[0].s_warehouse, fake.items[0].t_warehouse),
+			("EMP-MSL", "DEPT-RM"),
+		)
+		self.assertEqual(fake.items[0].qty, 1.0)
+		md = tree.material_details[0]
+		self.assertEqual(md.receive_qty, 3.0)
+		self.assertEqual(md.pending_qty, 0.0)
+		self.assertEqual(tree.status, "Received")
+
+	def test_casting_receive_pending_cap_blocks_over_receive(self):
+		# The (recv + loss) <= pending cap still fires for casting -> leftover-only, no double-count.
+		tree = self._casting_tree(issue=3.0, receive=2.0)  # pending 1
 		with self.assertRaises(ValidationError):
 			_run(
 				tse.receive_material,
 				tree,
-				[{"item_code": "GOLD-18KT", "receive_qty": 11.0}],
+				[{"item_code": "GOLD-18KT", "receive_qty": 2.0}],
 			)
+
+	def test_casting_receive_stamps_employee(self):
+		# The receive SE records the employee (tree.employee == the Issue Employee IR's employee).
+		tree = self._casting_tree(issue=3.0, receive=2.0)
+		fake = _run(
+			tse.receive_material, tree, [{"item_code": "GOLD-18KT", "receive_qty": 1.0}]
+		)[0]
+		self.assertEqual(fake.employee, tree.employee)
+
+	def test_casting_receive_legs_transfer_and_loss_repack(self):
+		tree = self._casting_tree(issue=2.0, receive=0.0)  # pending 2
+		fakes = _run(
+			tse.receive_material,
+			tree,
+			[{"item_code": "GOLD-18KT", "receive_qty": 1.0, "loss_qty": 1.0}],
+		)
+		self.assertEqual(len(fakes), 2)
+		se_recv, se_loss = fakes
+		# Received leg -> MAIN SLIP transfer, MSL -> Dept RM.
+		self.assertEqual(se_recv.stock_entry_type, "Material Transfer (MAIN SLIP)")
+		self.assertEqual(
+			(se_recv.items[0].s_warehouse, se_recv.items[0].t_warehouse),
+			("EMP-MSL", "DEPT-RM"),
+		)
+		# Loss leg -> Process Loss Repack, metal @ MSL consumed, ML variant @ Scrap produced.
+		self.assertEqual(se_loss.stock_entry_type, "Process Loss")
+		self.assertEqual(se_loss.employee, tree.employee)
+		consume, produce = se_loss.items
+		self.assertEqual((consume.s_warehouse, consume.t_warehouse), ("EMP-MSL", None))
+		self.assertEqual(produce.item_code, "GOLD-18KT-ML")
+		self.assertEqual(
+			(produce.s_warehouse, produce.t_warehouse), (None, "DEPT-SCRAP")
+		)
+		self.assertEqual(tree.status, "Received")
+
+	def test_casting_receive_partial_then_status(self):
+		tree = self._casting_tree(issue=3.0, receive=2.0)  # pending 1
+		_run(
+			tse.receive_material, tree, [{"item_code": "GOLD-18KT", "receive_qty": 0.5}]
+		)
+		self.assertEqual(tree.material_details[0].pending_qty, 0.5)
+		self.assertEqual(tree.status, "Partially Received")
+		_run(
+			tse.receive_material, tree, [{"item_code": "GOLD-18KT", "receive_qty": 0.5}]
+		)
+		self.assertEqual(tree.material_details[0].pending_qty, 0.0)
+		self.assertEqual(tree.status, "Received")
 
 	def test_is_casting_tree(self):
 		self.assertTrue(
@@ -1212,3 +1300,139 @@ class TestWarehouseResolvers(IntegrationTestCase):
 		with patch.object(tse.frappe.db, "get_all", return_value=dupes):
 			with self.assertRaises(ValidationError):
 				tse._resolve_scrap_warehouse("DEPT")
+
+	def test_source_warehouse_resolves_to_dept_rm(self):
+		# Default Issue source = the department Raw Material warehouse (not Manufacturing).
+		tree = SimpleNamespace(name="T1", department="DEPT", get=lambda k, d=None: None)
+		with patch.object(tse, "_get_department_rm_warehouse", return_value="DEPT-RM"):
+			self.assertEqual(tse._resolve_source_warehouse(tree), "DEPT-RM")
+
+	def test_source_warehouse_explicit_arg_wins(self):
+		# An explicit arg short-circuits before any tree/dept resolution.
+		tree = SimpleNamespace(
+			name="T1", department="DEPT", get=lambda k, d=None: "TREE-SRC"
+		)
+		self.assertEqual(
+			tse._resolve_source_warehouse(tree, "EXPLICIT-WH"), "EXPLICIT-WH"
+		)
+
+	def test_source_warehouse_tree_value_wins_over_dept(self):
+		# A source already stored on the tree wins over the dept-RM fallback.
+		tree = SimpleNamespace(
+			name="T1", department="DEPT", get=lambda k, d=None: "TREE-SRC"
+		)
+		self.assertEqual(tse._resolve_source_warehouse(tree), "TREE-SRC")
+
+
+# ---------------------------------------------------------------------------
+# Loss -> ML variant resolution
+# ---------------------------------------------------------------------------
+class TestLossItemResolution(IntegrationTestCase):
+	def test_resolve_loss_item_missing_variant_of_throws(self):
+		tree = SimpleNamespace(name="T1", company="_Test Company")
+		with patch.object(tse.frappe.db, "get_value", return_value=None):
+			with self.assertRaises(ValidationError):
+				tse._resolve_tree_loss_item(tree, "GOLD-18KT")
+
+	def test_resolve_loss_item_delegates_to_get_item_loss_item(self):
+		tree = SimpleNamespace(name="T1", company="_Test Company")
+		with (
+			patch.object(tse.frappe.db, "get_value", return_value="M"),
+			patch(
+				"jewellery_erpnext.jewellery_erpnext.doctype.main_slip.main_slip.get_item_loss_item",
+				return_value="GOLD-18KT-ML",
+			) as m,
+		):
+			out = tse._resolve_tree_loss_item(tree, "GOLD-18KT")
+		self.assertEqual(out, "GOLD-18KT-ML")
+		# variant_of derived from the Item; loss_type defaults to "Loss".
+		m.assert_called_once_with("_Test Company", "GOLD-18KT", "M", "Loss")
+
+
+# ---------------------------------------------------------------------------
+# Submit-to-lock (manual finalize)
+# ---------------------------------------------------------------------------
+class TestSubmitAndLock(IntegrationTestCase):
+	def _received_tree(self):
+		# Fully reconciled: issue 10 = receive 8 + loss 2, pending 0 -> _tree_status == "Received".
+		return _new_tree(
+			material_details=[
+				{
+					"item_code": "GOLD-18KT",
+					"issue_qty": 10.0,
+					"receive_qty": 8.0,
+					"loss_qty": 2.0,
+					"pending_qty": 0.0,
+				}
+			]
+		)
+
+	def test_submit_tree_sets_submitted(self):
+		# Fully received (pending 0): nothing to write off, so receive_material is never called.
+		tree = self._received_tree()
+		with (
+			patch.object(frappe, "has_permission", return_value=True),
+			patch.object(tse, "receive_material") as mock_recv,
+		):
+			tree.submit_tree()
+		mock_recv.assert_not_called()
+		self.assertEqual(tree.status, "Submitted")
+		tree.save.assert_called_once()
+
+	def test_submit_tree_writes_off_partial_then_locks(self):
+		# Partially Received (pending 7): submit books the leftover as loss via receive_material and
+		# then locks the tree at "Submitted".
+		tree = _new_tree(
+			material_details=[
+				{
+					"item_code": "GOLD-18KT",
+					"issue_qty": 10.0,
+					"receive_qty": 3.0,
+					"loss_qty": 0.0,
+					"pending_qty": 7.0,
+				}
+			]
+		)
+		with (
+			patch.object(frappe, "has_permission", return_value=True),
+			patch.object(tse, "receive_material") as mock_recv,
+		):
+			tree.submit_tree()
+		mock_recv.assert_called_once()
+		called_tree, called_rows = mock_recv.call_args[0]
+		self.assertIs(called_tree, tree)
+		self.assertEqual(called_rows, [{"item_code": "GOLD-18KT", "loss_qty": 7.0}])
+		self.assertEqual(tree.status, "Submitted")
+
+	def test_submit_tree_rejects_when_never_received(self):
+		# No receive activity at all -> _tree_status == "Issued" -> cannot submit (receive/reverse first).
+		tree = _new_tree(
+			material_details=[
+				{
+					"item_code": "GOLD-18KT",
+					"issue_qty": 10.0,
+					"receive_qty": 0.0,
+					"loss_qty": 0.0,
+					"pending_qty": 10.0,
+				}
+			]
+		)
+		with patch.object(frappe, "has_permission", return_value=True):
+			with self.assertRaises(ValidationError):
+				tree.submit_tree()
+
+	def test_submitted_tree_blocks_issue(self):
+		tree = self._received_tree()
+		tree.status = "Submitted"
+		with self.assertRaises(ValidationError):
+			_run(tse.issue_material, tree, "GOLD-18KT", 1.0)
+
+	def test_submitted_tree_blocks_receive(self):
+		tree = self._received_tree()
+		tree.status = "Submitted"
+		with self.assertRaises(ValidationError):
+			_run(
+				tse.receive_material,
+				tree,
+				[{"item_code": "GOLD-18KT", "receive_qty": 1.0}],
+			)
